+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Changed: single value refspec is now #abc.x to avoid #abc.h (height of #abc) conflicting
+  with #abc:h (position horizontally on the right of #abc). Mnemonic: '.' has a single dot
+  and returns a single numeric value...
+- Add support for locspecs in `points` attribute for polyline and polygon.
+- Add `dx`/`dy`/`dxy` support for `points` values, translating each point in turn.
+- Added `d-thin` and `d-biarrow` auto-style classes.
 - Add support for dirspec (h/H/v/V) relative positioning from a referenced element,
   e.g. `xy="#abc:h"`
 - Add various auto-style classes: `d-{colour}` & `d-fill-{colour}` for various colours,